+++ single-channel.php
@@ -23,7 +23,7 @@ if (have_posts()) {
         $dsp_api = new Dsp_External_Api_Request();
         $country_code = $dsp_api->get_country();
         $dspro_channel_geo = unserialize($channel_meta['dspro_channel_geo'][0]);
-        if($country_code && !in_array($country_code, $dspro_channel_geo) && !empty($dspro_channel_geo)){
+        if($country_code && !in_array("ALL", $dspro_channel_geo) && !in_array($country_code, $dspro_channel_geo) && !empty($dspro_channel_geo)){
             ?>
             <div class="custom-container container pb-5">
                 <div class="row no-gutters other-categories text-center">
@@ -372,7 +372,7 @@ if (!empty($trailer_id)) {
 
             <?php if (!empty($chnl_id)) { ?>
                 mountObj.channel_id = "<?php echo $chnl_id; ?>";
-                mountObj.channel_title = "<?php echo $chnl_title; ?>";
+                mountObj.channel_title = <?php echo json_encode($chnl_title); ?>;
             <?php } ?>
             <?php if(!empty($dspro_channel_id)) { ?>
                 mountObj.dspro_channel_id = "<?php echo $dspro_channel_id; ?>";